touch up

+++ src/hotspot/share/gc/shenandoah/shenandoahMmuTracker.cpp
@@ -140,7 +140,7 @@ void ShenandoahMmuTracker::record_mixed(uint gcid) {
   update_utilization(gcid, "Mixed Concurrent GC");
 }
 
-void ShenandoahMmuTracker::record_degenerated(size_t gcid, bool is_old_bootstrap) {
+void ShenandoahMmuTracker::record_degenerated(uint gcid, bool is_old_bootstrap) {
   if ((gcid == _most_recent_gcid) && _most_recent_is_full) {
     // Do nothing.  This is a redundant recording for the full gc that just completed.
   } else if (is_old_bootstrap) {
